+++ CHANGELOG.md
@@ -112,6 +112,7 @@
 - Enh #113: Add `HasMarkerUnits` trait and corresponding tests for managing SVG `markerUnits` attribute (@terabytesoftw)
 - Enh #114: Add `HasMarkerWidth` trait and corresponding tests for managing SVG `markerWidth` attribute (@terabytesoftw)
 - Enh #115: Add `HasOrient` trait and corresponding tests for managing SVG `orient` attribute (@terabytesoftw)
+- Enh #116: Add `HasRefX` trait and corresponding tests for managing SVG `refX` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024